updated shop models

+++ beekeeping_ecommerce/shop/views.py
@@ -1,5 +1,7 @@
+import random
+import string
+
 from datetime import datetime
-from urllib import request
 
 from django.shortcuts import render, get_object_or_404, redirect
 from django.views.generic import View
@@ -7,8 +9,17 @@ from django.core.exceptions import ObjectDoesNotExist
 from django.contrib import messages
 
 from django.urls import reverse
-from .models import Product, Order, OrderProduct, BillingAddress, Payment, Invoice
-from .forms import CheckoutForm
+from .models import (
+    Product,
+    Customer,
+    Order,
+    OrderProduct,
+    Address,
+    Payment,
+    Invoice,
+    Refund,
+)
+from .forms import CheckoutForm, RefundForm
 from beekeeping_ecommerce.contact.forms import DemoRequestForm
 from django.conf import settings
 
@@ -16,6 +27,60 @@ import stripe
 
 
 stripe.api_key = settings.STRIPE_API_SECRET_KEY
+
+
+def create_ref_code():
+    return "".join(random.choices(string.ascii_lowercase + string.digits, k=20))
+
+
+def get_customer(request):
+    """
+    This function is used to retrieve a customer.
+    It is used at each step of the purchasing process 
+    to get all info related to the order. If the customer
+    does not exists, an error message is sent and the user
+    redirected to the home page
+    """
+    try:
+        customer = Customer.objects.get(user_id=request.session.session_key)
+        return customer
+    except ObjectDoesNotExist:
+        messages.error(request, "Erreur d'identification de l'utilisateur. La procédure d'achat est interrompue")
+        return False
+
+
+def get_order_product(request, product):
+    # retrieve customer
+    customer = get_customer(request)
+
+    # check if the order_product exists else send an error message
+    try:
+        order_product = OrderProduct.objects.get(
+            customer=customer,
+            ordered=False,
+            product=product,
+        )
+        return order_product
+    except ObjectDoesNotExist:
+        messages.error(request, "Le produit n'existe pas dans le panier")
+        return False
+
+    
+
+def get_order(request):
+    # retrieve customer
+    customer = get_customer(request)
+
+    # check if the order exists else send an error message
+    try:
+        order = Order.objects.get(
+            customer=customer,
+            ordered=False,
+        )
+        return order
+    except ObjectDoesNotExist:
+        messages.error(request, "Le panier n'existe pas")
+        return False
 
 
 class HomeView(View):
@@ -28,18 +93,25 @@ class HomeView(View):
         form = DemoRequestForm(self.request.POST or None)
         if form.is_valid():
             demo_request = form.save()
-            context = {'demo_request': demo_request}
+            context = {"demo_request": demo_request}
             messages.info(self.request, "Votre demande a bien été prise en compte.")
             return render(self.request, "contact/confirmation.html", context)
         else:
             context = self.get_context_data(form)
-            messages.warning(self.request, "Le formulaire comporte des erreurs, merci de corriger et de le soumettre à nouveau.")
+            messages.warning(
+                self.request,
+                "Le formulaire comporte des erreurs, merci de corriger et de le soumettre à nouveau.",
+            )
             return render(self.request, "home.html", context)
 
     def get_context_data(self, form):
         try:
+            customer = Customer.objects.get(user_id=self.request.session.session_key)
+        except ObjectDoesNotExist:
+            customer = False
+        try:
             order = Order.objects.get(
-                user_id=self.request.session.session_key, ordered=False
+                customer=customer, ordered=False
             )
         except ObjectDoesNotExist:
             order = False
@@ -49,43 +121,25 @@ class HomeView(View):
 
 class CartView(View):
     def get(self, *args, **kwargs):
-        try:
-            order = Order.objects.get(
-                user_id=self.request.session.session_key, ordered=False
-            )
-        except ObjectDoesNotExist:
-            messages.info(self.request, "Votre panier est vide.")
-            return redirect("shop:home")
-        context = {"order": order, "products": order.products.all()}
+        order = get_order(self.request)
+        if not order:
+            return redirect("shop:home")        
+        context = {"order": order, "products": order.products.all().order_by('id')}
         return render(self.request, "shop/cart.html", context)
 
 
 class CheckoutView(View):
     def get(self, *args, **kwargs):
-        # check if a billing address has already been created for this session
-        billing_address_qs = BillingAddress.objects.filter(
-                user_id=self.request.session.session_key
-            )
-        if billing_address_qs.exists():
-            billing_address = billing_address_qs[0]
-        else:
-            billing_address = False
-        # Check if the order exists, otherwise return to homepage
-        try:
-            order = Order.objects.get(
-                user_id=self.request.session.session_key, ordered=False
-            )
-        except ObjectDoesNotExist:
-            messages.info(self.request, "Votre panier est vide.")
+        # Retrieve the customer and the order
+        order = get_order(self.request)
+        if not order:
             return redirect("shop:home")
-        # if the order container already a billing address this
-        # billing address is passed ot the form
-        if order.billing_address:
-            form = CheckoutForm(instance=order.billing_address)
-        # else the billing address related to the session is
-        # passed ot the form
-        elif billing_address:
-            form = CheckoutForm(instance=billing_address)
+
+        ## if the order contains already a billing address this
+        ## billing address is passed to the form
+        #if order.billing_address:
+        #    form = CheckoutForm(instance=order.billing_address)
+
         # else the form is blank
         else:
             form = CheckoutForm()
@@ -93,38 +147,53 @@ class CheckoutView(View):
         return render(self.request, "shop/checkout.html", context)
 
     def post(self, *args, **kwargs):
-        
-        # check if a billing address has already been created for this session
-        billing_address_qs = BillingAddress.objects.filter(
-                user_id=self.request.session.session_key
-            )
-        if billing_address_qs.exists():
-            billing_address = billing_address_qs[0]
-        else:
-            billing_address = False        
-        # Check if the order exists, otherwise return to homepage
-        try:
-            order = Order.objects.get(
-                user_id=self.request.session.session_key, ordered=False
-            )
-        except ObjectDoesNotExist:
-            messages.info(self.request, "Votre panier est vide.")
+
+        # get the order and the customer
+        customer = get_customer(self.request)
+        if not customer:
+            return redirect("shop:home")  
+        order = get_order(self.request)      
+        if not order:
             return redirect("shop:home")
-        # Check if the ordercontaines a billing address already
-        if order.billing_address:
-            billing_address = order.billing_address
-        # if a billing address exists from session or existing order, it is assigend to the form
-        if billing_address:
-            form = CheckoutForm(self.request.POST or None, instance=billing_address)
-        else:
-            form = CheckoutForm(self.request.POST or None)
+
+        form = CheckoutForm(self.request.POST or None)
         if form.is_valid():
             # create a new object or update the existing object
-            billing_address = form.save()
-            order.billing_address = billing_address
-            order.save()
+            billing_address = Address(
+                street_address=form.cleaned_data['street_address'],
+                street_address_line_2=form.cleaned_data['street_address_line_2'],
+                country=form.cleaned_data['country'],
+                zip_code=form.cleaned_data['zip_code'],
+                city=form.cleaned_data['city'],
+                customer=customer,
+                address_type='B',
+            )
+
+            # saves or updates the billing address and the order
+            if order.billing_address:
+                # updates the billing address if it already exists
+                billing_address.id = order.billing_address.pk
+                billing_address.save()
+            else:
+                # else create a new one
+                billing_address.save()
+                order.billing_address = billing_address
+                order.save()
+
+            # save the customer information
+            customer.first_name = form.cleaned_data['first_name']
+            customer.last_name = form.cleaned_data['last_name']
+            customer.company_name = form.cleaned_data['company_name']
+            customer.phone = form.cleaned_data['phone']
+            customer.email = form.cleaned_data['email']
+            customer.save()
+
+            # save the payment information and the same_address_information
+            order.payment_option = form.cleaned_data['payment_option']
+            order.same_shipping_address = form.cleaned_data['same_shipping_address']
+
             # check which payment option is selected to redirect to the appropriate view
-            if billing_address.payment_option == "S":
+            if order.payment_option == "S":
                 return redirect(
                     reverse("shop:payment", kwargs={"payment_option": "stripe"})
                 )
@@ -132,18 +201,18 @@ class CheckoutView(View):
                 messages.warning(
                     self.request, "Only Stripe payment is available right now"
                 )
-                return redirect("shop:checkout")
+                context = {"order": order, "form": form}
+                return render(self.request, "shop/checkout.html", context)
+        else:
+            context = {"order": order, "form": form}
+            return render(self.request, "shop/checkout.html", context)
 
 
 class PaymentView(View):
     def get(self, *args, **kwargs):
         # Check if the order exists, otherwise return to homepage
-        try:
-            order = Order.objects.get(
-                user_id=self.request.session.session_key, ordered=False
-            )
-        except ObjectDoesNotExist:
-            messages.info(self.request, "Votre panier est vide.")
+        order = get_order(self.request)
+        if not order:
             return redirect("shop:home")
 
         # Convert payment amount to cents for stripe processing
@@ -173,7 +242,7 @@ class PaymentView(View):
                 self.request, "Another problem occurred, maybe unrelated to Stripe."
             )
             return redirect("shop:checkout")
-
+        # else define context and render the view normally
         context = {
             "order": order,
             "client_secret": intent.client_secret,
@@ -188,13 +257,12 @@ class PaymentView(View):
 class StatusView(View):
     def get(self, *args, **kwargs):
         # Check if the order exists, otherwise return to homepage
-        try:
-            order = Order.objects.get(
-                user_id=self.request.session.session_key, ordered=False
-            )
-        except ObjectDoesNotExist:
-            messages.info(self.request, "Votre panier est vide.")
+        order = get_order(self.request)
+        if not order:
             return redirect("shop:home")
+
+        ## todo: mettre un check sur le paiement pour eviter que la page ne se
+        ## charge si l'order n'a pas été payé
 
         # Retrieve the payment intent ids returned by stripe after processing the payment
         payment_intent = self.request.GET.get("payment_intent")
@@ -210,13 +278,13 @@ class StatusView(View):
         payment = Payment.objects.create(
             stripe_intent_id=payment_intent,
             stripe_client_secret=payment_intent_client_secret,
-            user_id=self.request.session.session_key,
             amount=int(order.get_order_total_price() * 100),  # cents
             status=payment_status,
         )
 
         # Update the order with payment information, change ordered status to True
         order.payment = payment
+        order.ref_code = create_ref_code()
         order.ordered = True
         order.save()
         order_products = order.products.all()
@@ -261,18 +329,30 @@ class StatusView(View):
 
 
 def add_to_cart(request, slug):
+    # check is the product specified in the url exists
+    product = get_object_or_404(Product, slug=slug)
+
+    # if no session is active a session is created
     if not request.session.session_key:
         request.session.create()
-    product = get_object_or_404(Product, slug=slug)
+
+    # a customer is created or retrieved from the database
+    customer, is_customer_created = Customer.objects.get_or_create(
+        user_id=request.session.session_key
+    )
+
+    # get or create an order_product and an order
     order_product, is_order_product_created = OrderProduct.objects.get_or_create(
-        user_id=request.session.session_key,
+        customer=customer,
         ordered=False,
         product=product,
     )
     order, is_order_created = Order.objects.get_or_create(
-        user_id=request.session.session_key,
+        customer=customer,
         ordered=False,
     )
+
+    # update de cart
     if is_order_created or is_order_product_created:
         order.products.add(order_product)
         order.save()
@@ -285,51 +365,39 @@ def add_to_cart(request, slug):
 
 
 def remove_from_cart(request, slug):
+    # check is the product specified in the url exists
     product = get_object_or_404(Product, slug=slug)
-    try:
-        order_product = OrderProduct.objects.get(
-            user_id=request.session.session_key,
-            ordered=False,
-            product=product,
-        )
-    except ObjectDoesNotExist:
-        messages.info(request, "Le produit n'existe pas dans le panier")
-        return redirect("shop:cart")
-    try:
-        order = Order.objects.get(
-            user_id=request.session.session_key,
-            ordered=False,
-        )
-    except ObjectDoesNotExist:
-        messages.info(request, "Le panier n'existe pas")
+
+    order_product = get_order_product(request, product)
+    if not order_product:
         return redirect("shop:home")
+    order = get_order(request)
+    if not order:
+        return redirect("shop:home")
+    
+    # remove the product from the order
     order_product.delete()
     messages.info(request, "Le produit a été supprimé du panier")
+
+    # delete the order if it contains no more products
     if not order.products.all().exists():
         order.delete()
         messages.info(request, "Le panier a été supprimé")
+        return redirect("shop:home")
     return redirect("shop:cart")
 
 
 def remove_item_from_cart(request, slug):
+    # check is the product specified in the url exists
     product = get_object_or_404(Product, slug=slug)
-    try:
-        order_product = OrderProduct.objects.get(
-            user_id=request.session.session_key,
-            ordered=False,
-            product=product,
-        )
-    except ObjectDoesNotExist:
-        messages.info(request, "Le produit n'existe pas dans le panier")
-        return redirect("shop:cart")
-    try:
-        order = Order.objects.get(
-            user_id=request.session.session_key,
-            ordered=False,
-        )
-    except ObjectDoesNotExist:
-        messages.info(request, "Le panier n'existe pas")
+
+    order_product = get_order_product(request, product)
+    if not order_product:
         return redirect("shop:home")
+    order = get_order(request)
+    if not order:
+        return redirect("shop:home")
+
     if order_product.quantity == 1:
         order_product.delete()
         messages.info(request, "Le produit a été supprimé du panier")
@@ -340,4 +408,40 @@ def remove_item_from_cart(request, slug):
     if not order.products.all().exists():
         order.delete()
         messages.info(request, "Le panier a été supprimé")
+        return redirect("shop:home")
     return redirect("shop:cart")
+
+
+class RequestRefundView(View):
+    def get(self, *args, **kwargs):
+        form = RefundForm()
+        context = {"form": form}
+        return render(self.request, "shop/request_refund.html", context)
+
+    def post(self, *args, **kwargs):
+        form = RefundForm(self.request.POST)
+        if form.is_valid():
+            ref_code = form.cleaned_data.get("ref_code")
+            message = form.cleaned_data.get("message")
+            email = form.cleaned_data.get("email")
+
+            try:
+                # edit the order
+                order = Order.objects.get(ref_code=ref_code)
+                order.refund_requested = True
+                order.save()
+
+                # store the refund
+                Refund.objects.create(order=order, message=message, email=email)
+                messages.info(
+                    self.request,
+                    "Votre demande de remboursement a bien été prise en compte.",
+                )
+
+                return redirect("shop:request-refund")
+
+            except ObjectDoesNotExist:
+                messages.warning(self.request, "Ce numéro de commande n'existe pas.")
+
+            context = {"form": form}
+            return render(self.request, "shop/request_refund.html", context)
